move ron logic into bgm module
