+++ pm64/Cargo.toml
@@ -8,7 +8,9 @@ edition = "2018"
 lazy_static = "1"
 log = "0.4"
 midly = { version = "0.5", optional = true, default-features = false, features = ["std", "alloc"] }
+regex = "1.12.2"
 rmp-serde = "1.3.0"
+ron = "0.12.0"
 serde = "1"
 serde_derive = "1"
 typescript-type-def = "0.5"
